// src/generated/api/model/course.ts
/**
 * Registrar Backend API
 * Manages the Alloy Analyzer & the model in the background.
 *
 * The version of the OpenAPI document: 0.1.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Course { 
    code?: string;
    dept?: string;
    number?: number;
    name?: string;
    credits?: number;
    semester?: string;
    prereqs?: string;
    description?: string;
}